[CDEC-341] Refactor Pos.Core.Ssc module set
