+++ core/cardano-sl-core.cabal
@@ -112,9 +112,16 @@ library
                        Pos.Core.Slotting.Util
 
                        -- SSC
-                       Pos.Core.Ssc.Types
-                       Pos.Core.Ssc.Util
-                       Pos.Core.Ssc.Vss
+                       Pos.Core.Ssc.Commitment
+                       Pos.Core.Ssc.CommitmentsMap
+                       Pos.Core.Ssc.Opening
+                       Pos.Core.Ssc.OpeningsMap
+                       Pos.Core.Ssc.Payload
+                       Pos.Core.Ssc.Proof
+                       Pos.Core.Ssc.SharesDistribution
+                       Pos.Core.Ssc.SharesMap
+                       Pos.Core.Ssc.VssCertificate
+                       Pos.Core.Ssc.VssCertificatesMap
 
                        -- Update
                        Pos.Core.Update.Types
