+++ core/src/Pos/Core/Genesis/Canonical.hs
@@ -40,8 +40,8 @@ import           Pos.Core.Genesis.Types (GenesisAvvmBalances (..), GenesisData (
                                          GenesisVssCertificatesMap (..), GenesisWStakeholders (..))
 import           Pos.Core.ProtocolConstants (VssMaxTTL (..), VssMinTTL (..))
 import           Pos.Core.Slotting.Types (EpochIndex (..), Timestamp (..))
-import           Pos.Core.Ssc.Types (VssCertificate (..), VssCertificatesMap (..))
-import           Pos.Core.Ssc.Vss (validateVssCertificatesMap)
+import           Pos.Core.Ssc (VssCertificate (..), VssCertificatesMap (..),
+                               validateVssCertificatesMap)
 import           Pos.Core.Update.Types (BlockVersionData (..), SoftforkRule (..))
 import           Pos.Crypto (ProxyCert, ProxySecretKey (..), PublicKey, RedeemPublicKey, Signature,
                              decodeAbstractHash, fromAvvmPk, fullProxyCertHexF, fullPublicKeyF,
