+++ core/src/Pos/Core/Genesis/Types.hs
@@ -38,10 +38,9 @@ import           Serokell.Util (allDistinct, mapJson)
 import           Pos.Binary.Crypto ()
 import           Pos.Core.Common (Address, Coin, CoinPortion, SharedSeed, StakeholderId)
 import           Pos.Core.Delegation (ProxySKHeavy)
-import           Pos.Core.ProtocolConstants (ProtocolConstants (..),
-                                             VssMaxTTL (..), VssMinTTL (..))
+import           Pos.Core.ProtocolConstants (ProtocolConstants (..), VssMaxTTL (..), VssMinTTL (..))
 import           Pos.Core.Slotting.Types (Timestamp)
-import           Pos.Core.Ssc.Types (VssCertificatesMap, getVssCertificatesMap)
+import           Pos.Core.Ssc (VssCertificatesMap, getVssCertificatesMap)
 import           Pos.Core.Update.Types (BlockVersionData)
 import           Pos.Crypto.Configuration (ProtocolMagic)
 import           Pos.Crypto.Signing (RedeemPublicKey)
