+++ core/src/Pos/Core/Ssc.hs
@@ -1,8 +1,23 @@
 module Pos.Core.Ssc
-       ( module Pos.Core.Ssc.Types
-       , module Pos.Core.Ssc.Util
-       , module Pos.Core.Ssc.Vss
+       ( module Pos.Core.Ssc.Commitment
+       , module Pos.Core.Ssc.CommitmentsMap
+       , module Pos.Core.Ssc.Opening
+       , module Pos.Core.Ssc.OpeningsMap
+       , module Pos.Core.Ssc.Payload
+       , module Pos.Core.Ssc.Proof
+       , module Pos.Core.Ssc.SharesDistribution
+       , module Pos.Core.Ssc.SharesMap
+       , module Pos.Core.Ssc.VssCertificate
+       , module Pos.Core.Ssc.VssCertificatesMap
        ) where
-import           Pos.Core.Ssc.Types
-import           Pos.Core.Ssc.Util
-import           Pos.Core.Ssc.Vss
+
+import           Pos.Core.Ssc.Commitment
+import           Pos.Core.Ssc.CommitmentsMap
+import           Pos.Core.Ssc.Opening
+import           Pos.Core.Ssc.OpeningsMap
+import           Pos.Core.Ssc.Payload
+import           Pos.Core.Ssc.Proof
+import           Pos.Core.Ssc.SharesDistribution
+import           Pos.Core.Ssc.SharesMap
+import           Pos.Core.Ssc.VssCertificate
+import           Pos.Core.Ssc.VssCertificatesMap
